+++ marshmallow/fields.py
@@ -189,17 +189,19 @@ class Field(FieldABC):
         does not succeed.
         """
         errors = []
+        kwargs = {}
         for validator in self.validators:
             try:
                 if validator(value) is False:
                     self.fail('validator_failed')
             except ValidationError as err:
+                kwargs.update(err.kwargs)
                 if isinstance(err.messages, dict):
                     errors.append(err.messages)
                 else:
                     errors.extend(err.messages)
         if errors:
-            raise ValidationError(errors)
+            raise ValidationError(errors, **kwargs)
 
     # Hat tip to django-rest-framework.
     def fail(self, key, **kwargs):